% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{as_flextable.array}
\alias{as_flextable.array}
\title{as_flextable.array}
\usage{
\method{as_flextable}{array}(
  x,
  row.title = if (has_DNM[1L]) DNM[1L] else " ",
  hline_i = integer(0L),
  vline_j = integer(0L),
  ...
)
}
\arguments{
\item{x}{a \link[base]{array}}

\item{row.title}{\link[base]{character} scalar}

\item{hline_i, vline_j}{..}

\item{...}{..}
}
\value{
Function \link{as_flextable.array} returns a \link[flextable]{flextable}.
}
\description{
..
}
\examples{
as_flextable(VADeaths) # 'matrix'
as_flextable(occupationalStatus) # ?flextable:::as_flextable.table
as_flextable.array(occupationalStatus)

tryCatch(as_flextable(UCBAdmissions), error = identity) # ?flextable:::as_flextable.table
tryCatch(as_flextable.array(UCBAdmissions), error = identity)
}
