% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftable.R
\name{as_flextable.ftable}
\alias{as_flextable.ftable}
\title{Convert \link[stats]{ftable} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{ftable}(x, ...)
}
\arguments{
\item{x}{an \link[stats]{ftable}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{as_flextable.ftable} returns a \link[flextable]{flextable}.
}
\description{
Convert \link[stats]{ftable} to \link[flextable]{flextable}
}
\examples{
esoph0 = esoph[c('tobgp', 'ncases', 'ncontrols')]
(x0 = ftable(xtabs(cbind(ncases, ncontrols) ~ ., data = esoph0)))
x0 |> as_flextable()
(x = ftable(xtabs(cbind(ncases, ncontrols) ~ ., data = esoph)))
x |> as_flextable()
}
\keyword{internal}
