% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{as_flextable.matrix}
\alias{as_flextable.matrix}
\title{Convert a \link[base]{double} or \link[base]{character} \link[base]{matrix} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{matrix}(
  x,
  row.title = if (has_DNM[1L]) DNM[1L] else " ",
  hline_i = integer(0L),
  vline_j = integer(0L),
  ...
)
}
\arguments{
\item{x}{a \link[base]{double} or \link[base]{character} \link[base]{array}}

\item{row.title}{\link[base]{character} scalar}

\item{hline_i, vline_j}{..}

\item{...}{..}
}
\value{
Function \code{\link[=as_flextable.matrix]{as_flextable.matrix()}} returns a \link[flextable]{flextable}.
}
\description{
..
}
\keyword{internal}
