% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{as_flextable.noquote}
\alias{as_flextable.noquote}
\title{as_flextable.noquote}
\usage{
\method{as_flextable}{noquote}(x, ...)
}
\arguments{
\item{x}{a \link[base]{noquote} object}

\item{...}{additional parameters of S3 generic \link[flextable]{as_flextable}}
}
\value{
Function \code{\link[=as_flextable.noquote]{as_flextable.noquote()}} returns a \link[flextable]{flextable}.
}
\description{
as_flextable.noquote
}
\keyword{internal}
