% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust_.R
\name{as_flextable.p_adjust}
\alias{as_flextable.p_adjust}
\title{\link{as_flextable.p_adjust}}
\usage{
\method{as_flextable}{p_adjust}(x, ...)
}
\arguments{
\item{x}{\code{p_adjust} object}

\item{...}{..}
}
\description{
\link{as_flextable.p_adjust}
}
\keyword{internal}
