% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.htest.R
\name{as_flextable.pairwise.htest}
\alias{as_flextable.pairwise.htest}
\title{as_flextable.pairwise.htest}
\usage{
\method{as_flextable}{pairwise.htest}(x, row.title = x$method, ...)
}
\arguments{
\item{x}{a \code{'pairwise.htest'} object, as returned from \link[stats]{pairwise.t.test},
\link[stats]{pairwise.wilcox.test} or
\link[stats]{pairwise.prop.test}.}

\item{row.title, ...}{additional parameters of \link{as_flextable.array}}
}
\description{
as_flextable.pairwise.htest
}
\examples{
aq = within(airquality, expr = {
  Month = factor(Month, labels = month.abb[5:9])
})
(x = with(aq, pairwise.t.test(Ozone, Month, pool.sd = FALSE, p.adj = 'none')))
as_flextable(x)
}
