% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyHSD.R
\name{as_flextable.TukeyHSD}
\alias{as_flextable.TukeyHSD}
\title{Convert \link[stats]{TukeyHSD} into \link[flextable]{flextable}}
\usage{
\method{as_flextable}{TukeyHSD}(x, ...)
}
\arguments{
\item{x}{a \link[stats]{TukeyHSD}}

\item{...}{additional parameters, currently not in use}
}
\description{
Convert \link[stats]{TukeyHSD} into a \link[flextable]{flextable}.
}
\keyword{internal}
