% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtabs.R
\name{as_flextable.xtabs}
\alias{as_flextable.xtabs}
\title{Convert \link[stats]{xtabs} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{xtabs}(x, ...)
}
\arguments{
\item{x}{an \link[stats]{xtabs} object}

\item{...}{additional parameters of function \link{as_flextable.array}}
}
\description{
..
}
\details{
Prints zero-counts as \code{'.'}.
}
\note{
Without this S3 dispatch, will be dispatched to \code{flextable:::as_flextable.table}.
}
\examples{
# ?stats::xtabs
typeof(UCBAdmissions) # 'double'
(m = xtabs(Freq ~ Gender + Admit, data = as.data.frame(UCBAdmissions)))
typeof(m) # double
flextable:::as_flextable.table(m)
as_flextable(m)
}
