% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as_flextable_dataframe}
\alias{as_flextable_dataframe}
\title{Alternative Conversion from \link[base]{data.frame} to \link[flextable]{flextable}}
\usage{
as_flextable_dataframe(x, ...)
}
\arguments{
\item{x}{\link[base]{data.frame}}

\item{...}{..}
}
\value{
Function \link{as_flextable_dataframe} returns a \link[flextable]{flextable}.
}
\description{
..
}
\note{
Name clash \code{flextable:::as_flextable.data.frame}.
}
\keyword{internal}
