% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{flextable.tzh-package}
\alias{flextable.tzh}
\alias{flextable.tzh-package}
\title{flextable.tzh: flextable.tzh}
\description{
T. Zhan's extension to David Gohel's package 'flextable'.
}
\note{
haha
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
