% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{flextable_aov}
\alias{flextable_aov}
\alias{as_flextable.anova}
\alias{as_flextable.aov}
\alias{as_flextable.summary.aov}
\alias{as_flextable.aovlist}
\alias{as_flextable.summary.aovlist}
\title{Convert \link[stats]{aov}, \link[stats]{anova} Objects To \link[flextable]{flextable}}
\usage{
\method{as_flextable}{anova}(x, row.title = " ", ...)

\method{as_flextable}{aov}(x, ...)

\method{as_flextable}{summary.aov}(x, ...)

\method{as_flextable}{aovlist}(x, ...)

\method{as_flextable}{summary.aovlist}(x, ...)
}
\arguments{
\item{x}{\link[stats]{aov}, \link[stats]{anova}, or their \link[base]{summary} returns}

\item{row.title}{\link[base]{character} scalar}

\item{...}{additional parameters, currently not in use}
}
\description{
Convert \link[stats]{aov}, \link[stats]{anova}, or their \link[base]{summary} returns to \link[flextable]{flextable}.
}
\keyword{internal}
