% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit.R
\name{flextable_matchit}
\alias{flextable_matchit}
\alias{as_flextable.matchit}
\alias{as_flextable.summary.matchit}
\title{Convert \link[MatchIt]{matchit} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{matchit}(x, ...)

\method{as_flextable}{summary.matchit}(x, fmt = "\%.3f", ...)
}
\arguments{
\item{x}{a \link[MatchIt]{matchit} object}

\item{...}{additional parameters of function \code{MatchIt:::summary.matchit}}

\item{fmt}{\link[base]{character} scalar, see function \link[base]{sprintf}}
}
\value{
Functions \link{as_flextable.matchit} and \link{as_flextable.summary.matchit} both return a \link[flextable]{flextable}.
}
\description{
Convert \link[MatchIt]{matchit} to \link[flextable]{flextable}
}
\examples{
library(MatchIt)
# ?MatchIt::lalonde 
# `treat` is the treatment arm (0/1), `re78` is the outcome
m1 = matchit(treat ~ age+educ+race+nodegree+married+re74+re75, data = lalonde)
m2 = matchit(treat ~ age+educ+race+nodegree+married+re74+re75, data = lalonde, ratio = 2)
m1 |> as_flextable()
m1 |> as_flextable(addlvariables = 're78')
m1 |> Sprintf.matchit()
m1 |> summary(addlvariables = 're78') |> Sprintf.summary.matchit()
m2 |> as_flextable()
m2 |> as_flextable(addlvariables = 're78')

}
