% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format2flextable.R
\name{format2flextable}
\alias{format2flextable}
\title{format2flextable}
\usage{
format2flextable(x)
}
\arguments{
\item{x}{\link[base]{data.frame}}
}
\value{
Function \code{\link[=format2flextable]{format2flextable()}} returns a \link[base]{data.frame}
}
\description{
format2flextable
}
\keyword{internal}
