% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format4flextable.R
\name{format4flextable}
\alias{format4flextable}
\title{format4flextable}
\usage{
format4flextable(x)
}
\arguments{
\item{x}{\link[base]{data.frame}}
}
\value{
Function \code{\link[=format4flextable]{format4flextable()}} returns a \link[base]{data.frame}
}
\description{
format4flextable
}
\keyword{internal}
