% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_by.R
\name{highlight_by}
\alias{highlight_by}
\alias{highlight_by.data.frame}
\alias{highlight_by.flextable}
\title{highlight_by}
\usage{
highlight_by(x, ...)

\method{highlight_by}{data.frame}(x, ...)

\method{highlight_by}{flextable}(x, i, ...)
}
\arguments{
\item{x}{a \link[base]{data.frame} or \link[flextable]{flextable}}

\item{...}{additional parameters for function \link[flextable]{highlight}}

\item{i}{\link[base]{logical} \link[base]{expression} indicating rows to \link[flextable]{highlight}: missing values are taken as false.}
}
\description{
highlight_by
}
\examples{
datasets::attitude |>
 highlight_by(i = learning > 55 | raises > 65, j = 'rating')
 
datasets::airquality |>
 head(n = 10L) |> # with missing
 highlight_by(i = Ozone > 20, j = c('Month', 'Day'))
}
\keyword{internal}
