% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hline_by.R
\name{hline_by}
\alias{hline_by}
\title{hline_by}
\usage{
hline_by(x, i, ...)
}
\arguments{
\item{x}{a \link[flextable]{flextable}}

\item{i}{\link[stats]{formula}}

\item{...}{additional parameters, currently no use}
}
\description{
hline_by
}
\keyword{internal}
