% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyHSD.R
\name{md_.TukeyHSD}
\alias{md_.TukeyHSD}
\title{R Markdown Lines for \link[stats]{TukeyHSD}}
\usage{
\method{md_}{TukeyHSD}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[stats]{TukeyHSD}
}
\examples{
m = aov(breaks ~ wool + tension, data = warpbreaks)
library(rmd.tzh); list(
  'aov' = m,
  '`TukeyHSD`' = m |> TukeyHSD(which = 'tension', ordered = TRUE)
) |> render_(file = 'aov_etc')
}
\keyword{internal}
