% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit.R
\name{md_matchit}
\alias{md_matchit}
\alias{md_.matchit}
\alias{md_.summary.matchit}
\title{R Markdown Lines for \link[MatchIt]{matchit}}
\usage{
\method{md_}{matchit}(x, xnm, ...)

\method{md_}{summary.matchit}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \link[MatchIt]{matchit}
}
\examples{
library(MatchIt)
m = matchit(treat ~ age+educ+race+nodegree+married+re74+re75, data = lalonde)
list(
 '`matchit1`' = m,
 '`matchit2`' = m |> summary(addlvariables = 're78')
) |> rmd.tzh::render_(file = 'matchit')
}
\keyword{internal}
