% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit.R
\name{Sprintf_matchit}
\alias{Sprintf_matchit}
\alias{Sprintf.matchit}
\alias{Sprintf.summary.matchit}
\title{Description of \link[MatchIt]{matchit}}
\usage{
Sprintf.matchit(x)

Sprintf.summary.matchit(x)
}
\arguments{
\item{x}{\link[MatchIt]{matchit} object}
}
\description{
Description of \link[MatchIt]{matchit}
}
\keyword{internal}
