% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{as_flextable.array}
\alias{as_flextable.array}
\title{as_flextable.array}
\usage{
\method{as_flextable}{array}(
  x,
  row.title = if (has_DNM[1L]) DNM[1L] else " ",
  caption = attr(x, which = "caption", exact = TRUE),
  hl = integer(0L),
  vl = integer(0L),
  ...
)
}
\arguments{
\item{x}{a \link[base]{array}}

\item{row.title}{\link[base]{character} scalar}

\item{caption}{\link[base]{character} scalar}

\item{hl, vl}{..}

\item{...}{..}
}
\value{
Function \link{as_flextable.array} returns a \link[flextable]{flextable}.
}
\description{
..
}
\examples{
library(flextable)
as_flextable(VADeaths)
as_flextable(occupationalStatus) # ?flextable:::as_flextable.table
as_flextable.array(occupationalStatus)
}
