% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{as_flextable.dist}
\alias{as_flextable.dist}
\title{Convert \link[stats]{dist} into \link[flextable]{flextable}}
\usage{
\method{as_flextable}{dist}(x, ...)
}
\arguments{
\item{x}{\link[stats]{dist} object}

\item{...}{additional parameters of function \link{as_flextable.array}}
}
\value{
Function \link{as_flextable.dist} returns a \link[flextable]{flextable}.
}
\description{
..
}
\examples{
# ?stats::dist
as.dist((1 - cor(USJudgeRatings))/2) |> as_flextable()
}
