% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{as_flextable.matrix}
\alias{as_flextable.matrix}
\title{Convert a \link[base]{matrix} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{matrix}(x, row.title = if (has_DNM[1L]) DNM[1L] else " ", ...)
}
\arguments{
\item{x}{a \link[base]{matrix}}

\item{row.title}{\link[base]{character} scalar}

\item{...}{..}
}
\value{
Function \code{\link[=as_flextable.matrix]{as_flextable.matrix()}} returns a \link[flextable]{flextable}.
}
\description{
..
}
\keyword{internal}
