% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{as_flextable.noquote}
\alias{as_flextable.noquote}
\title{as_flextable.noquote}
\usage{
\method{as_flextable}{noquote}(x, ...)
}
\arguments{
\item{x}{a \link[base]{noquote} object}

\item{...}{additional parameters of function \link{as_flextable.array}}
}
\value{
Function \link{as_flextable.noquote} returns a \link[flextable]{flextable}.
}
\description{
as_flextable.noquote
}
