% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyHSD.R
\name{as_flextable.TukeyHSD}
\alias{as_flextable.TukeyHSD}
\title{Convert \link[stats]{TukeyHSD} into \link[flextable]{flextable}}
\usage{
\method{as_flextable}{TukeyHSD}(
  x,
  row.title = deparse1(attr(x, which = "orig.call", exact = TRUE)$formula[[2L]]),
  ...
)
}
\arguments{
\item{x}{a \link[stats]{TukeyHSD}}

\item{row.title}{\link[base]{character} scalar,
default is the left-hand-side of \link[stats]{aov} formula}

\item{...}{additional parameters, currently not in use}
}
\description{
Convert \link[stats]{TukeyHSD} into a \link[flextable]{flextable}.
}
\examples{
summary(fm1 <- aov(breaks ~ wool + tension, data = warpbreaks))
fm1 |> TukeyHSD(which = 'tension', ordered = TRUE) |> as_flextable()
fm1 |> TukeyHSD(which = 'tension') |> as_flextable()
fm1 |> TukeyHSD() |> as_flextable()
}
\keyword{internal}
