% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as_flextable_dataframe}
\alias{as_flextable_dataframe}
\title{Alternative Conversion from \link[base]{data.frame} to \link[flextable]{flextable}}
\usage{
as_flextable_dataframe(
  x,
  hline_i = integer(0L),
  vline_j = attr(x, which = "vline_j", exact = TRUE) \%||\% integer(0L),
  ...
)
}
\arguments{
\item{x}{\link[base]{data.frame}}

\item{hline_i}{..}

\item{vline_j}{..}

\item{...}{..}
}
\value{
Function \link{as_flextable_dataframe} returns a \link[flextable]{flextable}.
}
\description{
..
}
\note{
Name clash \code{flextable:::as_flextable.data.frame}.
}
\examples{
as_flextable_dataframe(warpbreaks, hline_i = ~ tension)

library(survival)
aml2 = within(aml, expr = {
  edp = Surv(time, status)
  time = status = NULL
})
as_flextable_dataframe(aml2, hline_i = ~ x)

}
\keyword{internal}
