% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable_dataframe}
\alias{as_flextable_dataframe}
\title{Alternative Conversion from \link[base]{data.frame} to \link[flextable]{flextable}}
\usage{
as_flextable_dataframe(
  x,
  caption = NULL,
  hline_i = integer(0L),
  vline_j = attr(x, which = "vline_j", exact = TRUE) \%||\% integer(0L),
  highlight_j = attr(x, which = "highlight_j", exact = TRUE) \%||\% integer(0L),
  ...
)
}
\arguments{
\item{x}{\link[base]{data.frame}}

\item{caption}{..}

\item{hline_i}{..}

\item{vline_j}{..}

\item{highlight_j}{..}

\item{...}{..}
}
\value{
Function \link{as_flextable_dataframe} returns a \link[flextable]{flextable}.
}
\description{
..
}
\note{
Name clash \code{flextable:::as_flextable.data.frame}.
}
\examples{
as_flextable_dataframe(warpbreaks, hline_i = ~ tension)
as_flextable_dataframe(swiss, highlight_j = 3L)

}
