% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{flextable_aov}
\alias{flextable_aov}
\alias{as_flextable.anova}
\alias{as_flextable.aov}
\alias{as_flextable.summary.aov}
\alias{as_flextable.aovlist}
\alias{as_flextable.summary.aovlist}
\title{Convert \link[stats]{aov}, \link[stats]{anova} Objects To \link[flextable]{flextable}}
\usage{
\method{as_flextable}{anova}(x, fmt = "\%.3f", row.title = " ", ...)

\method{as_flextable}{aov}(x, ...)

\method{as_flextable}{summary.aov}(x, ...)

\method{as_flextable}{aovlist}(x, ...)

\method{as_flextable}{summary.aovlist}(x, ...)
}
\arguments{
\item{x}{\link[stats]{aov}, \link[stats]{anova}, or their \link[base]{summary} returns}

\item{fmt}{\link[base]{character} scalar, see function \link[base]{sprintf}}

\item{row.title}{\link[base]{character} scalar}

\item{...}{additional parameters, currently not in use}
}
\description{
Convert \link[stats]{aov}, \link[stats]{anova}, or their \link[base]{summary} returns to \link[flextable]{flextable}.
}
\examples{
# ?stats::aov
op = options(contrasts = c("contr.helmert", "contr.poly"))
class(m1 <- aov(yield ~ block + N*P*K, data = npk)) # 'aov'
as_flextable(m1)
aov(yield ~ block + N * P + K, npk) |> as_flextable()
class(m2 <- aov(yield ~  N*P*K + Error(block), data = npk)) # 'aovlist'
as_flextable(m2)
}
\keyword{internal}
