% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchit.R
\name{flextable_matchit}
\alias{flextable_matchit}
\alias{as_flextable.matchit}
\alias{as_flextable.summary.matchit}
\title{Convert \link[MatchIt]{matchit} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{matchit}(x, ...)

\method{as_flextable}{summary.matchit}(x, ...)
}
\arguments{
\item{x}{a \link[MatchIt]{matchit} object}

\item{...}{additional parameters of function \code{MatchIt:::summary.matchit}}
}
\value{
Functions \code{\link[=as_flextable.matchit]{as_flextable.matchit()}} and \code{\link[=as_flextable.summary.matchit]{as_flextable.summary.matchit()}} both return a \link[flextable]{flextable}.
}
\description{
Convert \link[MatchIt]{matchit} to \link[flextable]{flextable}
}
\keyword{internal}
