% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format2flextable.R
\name{format2flextable}
\alias{format2flextable}
\title{format2flextable}
\usage{
format2flextable(x)
}
\arguments{
\item{x}{\link[base]{data.frame}}
}
\value{
Function \link{format2flextable} returns a \link[base]{data.frame}
}
\description{
format2flextable
}
\examples{
library(survival)
aml2 = within(aml, expr = {
  edp = Surv(time, status)
  time = status = NULL
})

# ?flextable:::as_flextable.data.frame cannot process 'matrix' columns, as of 2024-10-27
tryCatch(as_flextable(aml2), error = identity)

aml2 |> format2flextable() |> as_flextable()
}
