% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.pairwise.htest.R
\name{md_.pairwise.htest}
\alias{md_.pairwise.htest}
\title{Markdown Lines for \code{pairwise.htest} object}
\usage{
\method{md_}{pairwise.htest}(x, xnm, ...)
}
\arguments{
\item{x}{\code{pairwise.htest} object}

\item{xnm}{..}

\item{...}{..}
}
\description{
Markdown Lines for \code{pairwise.htest} object
}
\examples{
library(rmd.tzh); list(
  '`pairwise.htest`' = airquality |> 
    within.data.frame(expr = {
      Month = factor(Month, labels = month.abb[5:9])
    }) |>
    with(expr = pairwise.t.test(Ozone, Month, pool.sd = FALSE, p.adj = 'none'))
) |> render_(file = 'pairwise.htest')

}
