% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust_.R
\name{p_adjust_}
\alias{p_adjust_}
\alias{p_adjust_.numeric}
\alias{p_adjust_.pairwise.htest}
\title{Adjusted \eqn{p}-Values in All Correction Methods}
\usage{
p_adjust_(x)

\method{p_adjust_}{numeric}(x)

\method{p_adjust_}{pairwise.htest}(x)
}
\arguments{
\item{x}{see \strong{Usage}}
}
\value{
All method dispatches of generic function \link{p_adjust_} return a \link[base]{matrix} of adjusted \eqn{p}-values using all available \link[stats]{p.adjust.methods}.
}
\description{
Adjusted \eqn{p}-values in all correction methods.
}
\details{
Function \link{p_adjust_.numeric} speeds up function \link[stats]{p.adjust}.
}
\examples{
\donttest{for (i in seq_len(10L)) {
 n = sample(20:30, size = 1L)
 p = runif(n, min = .Machine$double.eps, max = 1 - .Machine$double.eps)
 (tmp = p_adjust_(p))
 stopifnot(identical(tmp, cbind(
  holm = p.adjust(p, 'holm'),
  hochberg = p.adjust(p, 'hochberg'),
  hommel = p.adjust(p, 'hommel'),
  bonferroni = p.adjust(p, 'bonferroni'),
  BH = p.adjust(p, 'BH'),
  BY = p.adjust(p, 'BY'),
  none = p)))
}}
}
