% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_.R
\name{subset_}
\alias{subset_}
\title{Inspect a Subset of \link[base]{data.frame}}
\usage{
subset_(x, subset, select, select_pattern, avoid, avoid_pattern)
}
\arguments{
\item{x}{a \link[base]{data.frame}}

\item{subset}{\link[base]{logical} \link[base]{expression},
see function \link[base]{subset.data.frame}}

\item{select}{\link[base]{character} \link[base]{vector},
columns to be selected,
see function \link[base]{subset.data.frame}}

\item{select_pattern}{regular expression \link[base]{regex}
for multiple columns to be selected}

\item{avoid}{\link[base]{character} \link[base]{vector},
columns to be avoided}

\item{avoid_pattern}{regular expression \link[base]{regex},
for multiple columns to be avoided}
}
\value{
Function \code{\link[=subset_]{subset_()}} returns a \link[flextable]{flextable}.
}
\description{
..
}
\details{
Function \code{\link[=subset_]{subset_()}} is different from function
\link[base]{subset.data.frame}, such that
\itemize{
\item {if both \code{select} and \code{select_pattern} are missing, only variables mentioned in \code{subset} are selected;}
\item {be able to select all variables, except those in \code{avoid} and \code{avoid_pattern}.}
}
}
\examples{
subset_(trees, Girth > 9 & Height < 70)
subset_(swiss, Fertility > 80, avoid = 'Catholic')
subset_(warpbreaks, wool == 'K')

}
\keyword{internal}
