% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.R
\name{symmetric2flextable}
\alias{symmetric2flextable}
\title{Convert \link[base]{isSymmetric} Matrix to \link[flextable]{flextable}}
\usage{
symmetric2flextable(x, ...)
}
\arguments{
\item{x}{\link[base]{isSymmetric} \link[base]{matrix}}

\item{...}{additional parameters of function \link[base]{format.default}}
}
\value{
Function \code{\link[=symmetric2flextable]{symmetric2flextable()}} returns a \link[flextable]{flextable}.
}
\description{
To convert an \link[base]{isSymmetric} matrix to a \link[flextable]{flextable}.
}
\seealso{
\link[stats]{cov2cor}
}
\keyword{internal}
